uint16-map everything

+++ muddersMIMA_firmware/2d_map.h
@@ -3,6 +3,6 @@
 #ifndef twod_map_h
 	#define twod_map_h
 
-	uint8_t evaluate_2d_map(uint8_t* map, uint8_t* dim1_map, uint8_t dim1_size, uint8_t dim1, uint8_t* dim2_map, uint8_t dim2_size, uint8_t dim2);
-    uint8_t two_d_map_getLastUsedMapIndex();
+	uint16_t evaluate_2d_map(uint16_t* map, uint16_t* dim1_map, uint8_t dim1_size, uint16_t dim1, uint16_t* dim2_map, uint8_t dim2_size, uint16_t dim2);
+    uint16_t two_d_map_getLastUsedMapIndex();
 #endif
